Refactor fixed-width integer types into its own module
