+++ src/hermit/mod.rs
@@ -13,15 +13,6 @@
 // Ported by Colin Fink <colin.finck@rwth-aachen.de>
 //       and Stefan Lankes <slankes@eonerc.rwth-aachen.de>
 
-pub type int8_t = i8;
-pub type int16_t = i16;
-pub type int32_t = i32;
-pub type int64_t = i64;
-pub type uint8_t = u8;
-pub type uint16_t = u16;
-pub type uint32_t = u32;
-pub type uint64_t = u64;
-
 pub type c_schar = i8;
 pub type c_uchar = u8;
 pub type c_short = i16;
