+++ src/wasi.rs
@@ -19,14 +19,6 @@ pub type intptr_t = isize;
 pub type uintptr_t = usize;
 pub type off_t = i64;
 pub type pid_t = i32;
-pub type int8_t = i8;
-pub type uint8_t = u8;
-pub type int16_t = i16;
-pub type uint16_t = u16;
-pub type int32_t = i32;
-pub type uint32_t = u32;
-pub type int64_t = i64;
-pub type uint64_t = u64;
 pub type clock_t = c_longlong;
 pub type time_t = c_longlong;
 pub type c_double = f64;
